Do not crash the server on client timeout
If there is an interest on crashing the server when
the client dies, the developer should rather link those
two processes. Which is in fact recommended if you are
running anything that requires multiple commands, like
transactions.

Otherwise, the connection behaves as any other GenServer
that just perform operations in the queue, even if the
client died.