Refactor Connection.query!/4 and Connection.unlisten!/3
Both those function basically replicated the behaviour of their non-bang
counterparts. While this may be a tiny bit more efficient than building on the
non-bang version, it duplicates a lot of code (almost all code in the functions)
and is therefore slightly hard to read and more error prone. This commit changes
this situation by building `Postgrex.Connection.query!/4` on top of
`Postgrex.Connection.query/4` and `Postgrex.Connection.unlisten!/3` on top of
`Postgrex.Connection.unlisten/3`.