Add a type server for caching pg types
Previously each connection kept about 100kb of type info.
With this patch, they are now shared and stored in ETS,
which is used directly when encoding/decoding data.

This will also allow us to move the encoding/decoding to
the client, allowing us to get better concurrency from
our connections.