Add `:decode_binary` option to copy decoded values
`:decode_binary` is either `:copy` or `:reference`.

* `:copy` uses `:binary.copy/1` to copy values decoded as binaries.
* `:reference` uses a reference counted binary from the binary streamed from the
socket.

If binaries are being held on to for a substantial period of time, sent to
other process, or inserted into ETS it may prevent garbage collection of a
larger binary. This option copies by default so that extra memory is not
consumed. If binaries are going to be garbage collected soon it will be
more efficient to use a reference binary.