Use types table lock when rebootstrapping
Avoid doing multiple rebootstraps when multiple connections connect
concurrently. We only need to do this once because the connections
would get the same results. Therefore only the first connection
reboostraps, and others wait for it to finish. If it fails the next
connection will try, just like the initial bootstrap.

Similarly on initial bootstrap we skip rebootstrapping on connections
that were waiting for initial bootstrap.