Move encode to own function and decode dispatch to case
If encode in function and then inline with the @compile the generated
code is the same but the file/line numbers match up.

Moving the decode dispatch to a case avoids the binary match for each
value in a row. The "pointless" matches were occuring as part of the
continuous binary match context optimisation.