Simplify query in bootstrap_query method
Previously, bootstrap_query constructed a pretty complicated `SELECT`
statement that used a subquery, array and generate_series to exclude
rows whose oids were part of a predefined set. The query looked like
this:

```
SELECT ... WHERE t.oid NOT IN
  (SELECT (ARRAY[integer, literals, ...])[i]
   FROM generate_series(1, length_of_above_array))
```

This subquery is now replaced by a simpler list of integer literals,
which should be more performant and easier to understand.

Now, the query looks like this:

```
SELECT ... WHERE t.oid NOT IN (integer, literals, ...)
```